util: use intsets.Sparse in FastIntSet
Improving the code for FastIntSet to use a `intsets.Sparse` when we have
elements outside of the [0,64) range. We don't just use `Sparse` directly
because it's larger (56 bytes) and must be on the heap (we can't have slices of
`Sparse`).

Also improving the interface to use `int`s and adding a `SubsetOf` method.

+++ pkg/sql/plan_ordering.go
@@ -79,7 +79,7 @@ func sortOrdering(n *sortNode) orderingInfo {
 		ord.ordering = make(sqlbase.ColumnOrdering, 0, len(n.ordering))
 		for _, o := range n.ordering {
 			// Skip any constant columns (we preserve them below).
-			if !underlying.constantCols.Contains(uint32(o.ColIdx)) {
+			if !underlying.constantCols.Contains(o.ColIdx) {
 				ord.addOrderColumn(o.ColIdx, o.Direction)
 			}
 		}
